requirements.txt
# requirements.txt

aiohttp==3.9.0
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
discord.py==2.3.2
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
multidict==6.0.4
notion-client==2.1.0
sniffio==1.3.0
yarl==1.9.3
